% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{col_identifiers_example}
\alias{col_identifiers_example}
\title{Example Column Identifiers}
\format{
A character vector with 1 element.
}
\usage{
col_identifiers_example
}
\description{
A character vector of example column identifiers used in the \code{rowcolr} package.
}
\examples{
col_identifiers_example
}
\keyword{datasets}
