% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{row_identifiers_example}
\alias{row_identifiers_example}
\title{Example Row Identifiers}
\format{
A character vector with 23 elements.
}
\usage{
row_identifiers_example
}
\description{
A character vector of example row identifiers used in the \code{rowcolr} package.
}
\examples{
row_identifiers_example
}
\keyword{datasets}
