% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowcolr_example.R
\name{rowcolr_example}
\alias{rowcolr_example}
\title{Get the path to an example Excel file}
\usage{
rowcolr_example(path = NULL)
}
\arguments{
\item{path}{Optional. A specific file name within the \code{extdata} directory. Defaults to \code{NULL}, which lists available files.}
}
\value{
A character string with the file path or a directory listing.
}
\description{
Retrieves the path to the example \code{example.xlsx} file included in the \code{rowcolr} package.
}
